backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
